+++ TietoCalorieApp/Models/DTO/AddFoodDTO.cs
@@ -6,9 +6,9 @@
         public string Name { get; set; }
         public int Weight { get; set; }
         public int CalorieCount { get; set; }
-        public int CarbsCount { get; set; }
-        public int ProteinCount { get; set; }
-        public int FatsCount { get; set; }
+        public double CarbsCount { get; set; }
+        public double ProteinCount { get; set; }
+        public double FatsCount { get; set; }
 
     }
 }
