+++ TietoCalorieApp/Models/Food.cs
@@ -4,9 +4,9 @@ public class Food
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public int Weight { get; set; }
+    public double Weight { get; set; }
     public int CalorieCount { get; set; }
-    public int CarbsCount { get; set; }
-    public int ProteinCount { get; set; }
-    public int FatsCount { get; set; }
+    public double CarbsCount { get; set; }
+    public double ProteinCount { get; set; }
+    public double FatsCount { get; set; }
 }
